[ALPS03385589] wlan: Fix Coverity defect: 2357654
[Detail]
prRegInfo->cTxBackOffMaxPower5G with type UINT8 (8 bits, unsigned) is promoted
in prRegInfo->cTxBackOffMaxPower5G << 24 to type int (32 bits, signed), then
sign-extended to type unsigned long (64 bits, unsigned).
If prRegInfo->cTxBackOffMaxPower5G << 24 is greater than 0x7FFFFFFF,
the upper bits of the result will all be 1 and the value will be changed a lot.

[Solution]
use type casting to change cTxBackOffMaxPower5G from a 8-bits data to a 64-bit
data before left shifting.

MTK-Commit-Id: 8e7dd5c0fb0f6650b287ae5a7d639bc5f5a129f0

CR-Id: ALPS03385589
Feature: Wifi Driver

+++ drivers/misc/mediatek/connectivity/wlan/gen2/os/linux/gl_wext_priv.c
@@ -3316,7 +3316,7 @@ _priv_set_struct(IN struct net_device *prNetDev,
 					TxPwrBackOffParam |= prRegInfo->bTxPowerLimitEnable2G;
 					TxPwrBackOffParam |= prRegInfo->cTxBackOffMaxPower2G << 8;
 					TxPwrBackOffParam |= prRegInfo->bTxPowerLimitEnable5G << 16;
-					TxPwrBackOffParam |= prRegInfo->cTxBackOffMaxPower5G << 24;
+					TxPwrBackOffParam |= (ULONG)prRegInfo->cTxBackOffMaxPower5G << 24;
 					rStatus = nicTxPowerBackOff(prGlueInfo->prAdapter, TxPwrBackOffParam);
 				} else {
 					DBGLOG(REQ, INFO, "Stop BackOff\n");
